requirements.txt
Flask
openai-whisper
ffmpeg-python 